REFACTOR: Rename package from panos-cli to pan-os-cli
