+++ pyproject.toml
@@ -1,10 +1,10 @@
 [tool.poetry]
-name = "panos-cli"
+name = "pan-os-cli"
 version = "0.1.0"
 description = "CLI tool for managing PAN-OS configurations with multi-threading support"
 authors = ["Calvin Remsburg <cremsburg@protonmail.com>"]
 readme = "README.md"
-packages = [{include = "panos_cli", from = "src"}]
+packages = [{include = "pan_os_cli", from = "src"}]
 
 [tool.poetry.dependencies]
 python = ">=3.10,<3.13"
@@ -28,7 +28,7 @@ flake8 = "^7.2.0"
 yamllint = "^1.37.0"
 
 [tool.poetry.scripts]
-panos-cli = "panos_cli.main:app"
+pan-os-cli = "pan_os_cli.main:app"
 
 [tool.pytest]
 testpaths = ["tests"]
