+++ src/pan_os_cli/commands/__init__.py
@@ -1,0 +1,5 @@
+"""Command modules for pan-os-cli."""
+
+from pan_os_cli.commands import objects
+
+__all__ = ["objects"]
