+++ src/pan_os_cli/models/objects.py
