--- src/panos_cli/commands/__init__.py
@@ -1,5 +1,0 @@
-"""Command modules for panos-cli."""
-
-from panos_cli.commands import objects
-
-__all__ = ["objects"]
